add VariantSourceEntity
Note that the collection name is specified through the injection
of a bean, so other projects like eva-ws can modify it. Example:

```
    @Bean
    public String mongoCollectionsFiles() {
        return "files_1_2";
    }
```
It's not necessary to hardcode the string there, so further
injection from properties can be done as well.

**Important**: for the `@Document` injection to work, it's
mandatory to add the ApplicationContext to the MongoMappingContext.
see http://stackoverflow.com/questions/41698445/cannot-resolve-bean-in-spel-for-spring-data-mongodb-collection-name?rq=1