bump again to 0.7.2
the reason to jump to 0.7.2 is that there was some confusion around
v0.7.1, which was attempted to be used, but was not deployed. However
there was some application like eva-accession that used 0.7.1.

To clear all misunderstandings, 0.7.2 is the newest at this moment and
is the one that should be used in every place.